pyproject.toml
[tool.poetry]
name = "RPN calculator"
version = "1.0.0"
description = ""
authors = ["Kseniia Kuzina <kse.kuzina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

Flask = "^3.0.3"
alembic = "^1.13.3"
SQLAlchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-restx = "^1.3.0"
pytest-flask = "^1.3.0"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"